🛠️ Corrige Jenkinsfile et pom.xml, suppression des conflits et amélioration du pipeline

+++ devops-fullstack/backend/backendDevops/src/main/java/tn/esprit/backendDevops/repository/HeuresSupRepository.java
@@ -1,7 +1,5 @@
 package tn.esprit.backendDevops.repository;
 
-
-
 import tn.esprit.backendDevops.entities.HeuresSup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -12,5 +10,6 @@ import java.util.List;
 @Repository
 public interface HeuresSupRepository extends JpaRepository<HeuresSup, Long> {
     // Méthode pour trouver les heures supplémentaires par employé et par plage de dates
-    List<HeuresSup> findByEmployeIdAndDateBetween(int employeId, LocalDate startDate, LocalDate endDate);
+    // Correction : Utilisation de Long pour employeId pour correspondre à l'ID de Employe
+    List<HeuresSup> findByEmployeIdAndDateBetween(Long employeId, LocalDate startDate, LocalDate endDate);
 }
